+++ envs/metadata.yaml
@@ -1,9 +1,8 @@
-name: default
+name: metadata
 channels:
     - conda-forge
 dependencies:
     - python=3.6
     - pyyaml=5.1
     - pip:
-        - pytest-pythonpath # adds the root path "./" to PYTHONPATH
         - -e ../../lib
