+++ lib/renewablepotentialslib/eligibility.py
@@ -1,4 +1,7 @@
 from enum import IntEnum, Enum
+import numpy as np
+
+DATATYPE = np.uint8
 
 class Eligibility(IntEnum):
     """Categories defining land eligibility for renewable power."""
@@ -123,3 +126,36 @@ class Potential(Enum):
 
     def __str__(self):
         return self.__repr__()
+
+
+def determine_eligibility(
+    land_cover, slope, bathymetry, building_share, urban_green_share,
+    max_slope, max_building_share, max_urban_green_share, max_depth_offshore
+):
+    # parameters
+    max_slope_pv = max_slope["pv"]
+    max_slope_wind = max_slope["wind"]
+    assert max_slope_pv <= max_slope_wind # wind can be built whereever pv can be built
+
+    # prepare masks
+    settlements = (building_share > max_building_share) | (urban_green_share > max_urban_green_share)
+    farm = np.isin(land_cover, FARM)
+    forest = np.isin(land_cover, FOREST)
+    other = np.isin(land_cover, OTHER)
+    water = np.isin(land_cover, WATER)
+    pv = (slope <= max_slope_pv) & ~settlements & (farm | other)
+    wind = (slope <= max_slope_wind) & ~settlements & (farm | forest | other)
+    offshore = (bathymetry > max_depth_offshore) & ~settlements & water
+
+    # allocate eligibility
+    land = np.ones_like(land_cover, dtype=DATATYPE) * Eligibility.NOT_ELIGIBLE
+    _add_eligibility(land, Eligibility.ROOFTOP_PV, settlements)
+    _add_eligibility(land, Eligibility.ONSHORE_WIND_AND_PV, wind & pv)
+    _add_eligibility(land, Eligibility.ONSHORE_WIND, wind & ~pv)
+    _add_eligibility(land, Eligibility.OFFSHORE_WIND, offshore)
+    return land
+
+
+def _add_eligibility(land, eligibility, mask):
+    assert all(land[mask] == Eligibility.NOT_ELIGIBLE), f"Overwriting other eligibility with {eligibility}."
+    land[mask] = eligibility
