+++ lib/setup.py
@@ -7,9 +7,19 @@ setup(
     version='0.1.0', # additionally defined in __init__.py
     description='Library code of the solar-and-wind-potentials workflow.',
     maintainer='calliope-project',
-    maintainer_email='bryn.pickering@usys.ethz.ch',
+    maintainer_email='tim.troendle@usys.ethz.ch',
     packages=find_packages(exclude=['tests*']),
     include_package_data=True,
+    install_requires=[
+        "pycountry",
+        "gdal",
+        "rasterio",
+        "geopandas",
+        "numpy",
+        "pandas",
+        "shapely",
+        "pyproj"
+    ],
     classifiers=[
         'Environment :: Console',
         'Intended Audience :: Science/Research',
