+++ tests/test_administrative_borders.py
@@ -1,4 +1,3 @@
-
 from pathlib import Path
 import os
